Added demo
Added a demo page demonstrating several use cases for the library. It is a
supplement to (rather than a substitute for) a proper README, which is still
lacking and will have to come in another commit.

Once merged, this should be pushed to the gh-pages branch on Github in order to
create a [project page][1] from which the demo can be served.

[1]: https://help.github.com/articles/creating-project-pages-manually

Change-Id: I7f6dab27c1b0d7e8b65f3730ce73a238557fc2b4
Reviewed-on: https://gerrit.causes.com/22228
Reviewed-by: Preston Guillory <preston.guillory@causes.com>
Tested-by: Preston Guillory <preston.guillory@causes.com>