Add Primer-like click deferral
One interesting aspect of Facebook's Primer library is that it runs
*above* the page content. If the user interacts with content while the
page is loading, that content is already initialized, whether or not all
the scripts (typically below the content) have run.

To see the problem, go to causes.com and find an element for which we
have some javascript hook, like an "Invite friends" link. It's supposed
to show a modal when you click it. But if you click it *right* after the
page becomes visible, no modal appears, and a hash gets appended to the
document's location. Another example is the tabs on an issues page. If
you click one too soon after the page appears, the browser follows it
like a normal link, rather than doing a partial replacement. It still
works, but the experience is sub-optimal.

This commit provides a hook that catches click events during page load
and re-dispatches them after our click handlers are set up. This
"primer" should be placed in a script tag in the document's head. It
operates only Replace-enhanced links -- not on forms, and not on
anything that doesn't use Replace. The minified version is just 245
bytes, so hopefully the impact on performance is small and we reap a net
benefit to the quality of user experience.

Change-Id: I28f91e1236c10b94dabe980c19612bc63cd7090c
Reviewed-on: https://gerrit.causes.com/22235
Tested-by: Preston Guillory <preston.guillory@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>