Import initializeEach() from causes.git
In causes.git we use initializeEach() to squeeze more value out of
replace.js. This commit imports initializeEach() into replace.js so that
others can benefit from the functionality. This is a straight move with
no modifications; it corresponds to initializeEach as it existed at the
current HEAD of the master branch in causes.git (commit
84912bc74c3046887b4).

A subsequent commit will do the same for our "data-lazy-url"
functionality, and at that point we'll be able to look at making
ajaxReplace an internal-only function (this will require some work on
the Causes side).

Related issue:
  https://github.com/causes/replacejs/issues/2
  ("Move lazy-load functionality into replacejs repo")

Change-Id: I18126382149f30c2917b527a03bafe0443d10db8
Reviewed-on: https://gerrit.causes.com/25809
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>