Import "data-lazy-url" functionality from causes.git
Much as we did in the last commit, here we're bringing across useful
functionality from causes.git (corresponding to commit
84912bc74c3046887b4 of the master branch); this time it is our
"data-lazy-url" functionality.

This is a straight move, and brings with it a dependency on the jQuery
waypoints plug-in:

  http://imakewebthings.com/jquery-waypoints/

In a subsequent commit, I'll turn this into a soft dependency (ie. the
behavior should degrade gracefully in the absence of the plug-in), but
for this initial commit I wanted this to be a clear-cut move operation
with no modifications.

Related issue:
  https://github.com/causes/replacejs/issues/2
  ("Move lazy-load functionality into replacejs repo")

Change-Id: Ia1a0e5213c7b6ec51ce08f4986b03f0815beb1df
Reviewed-on: https://gerrit.causes.com/25810
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>