Add bower.json
This will enable replace.js to be installed with Bower[0], which is
something we've recently started doing for our front-end dependencies
(for example, method-proxy-js[1]).

I'll be pushing a 0.0.1 tag shortly, and registering the package with
the Bower registry.

One point to note: jquery is a hard dependency and jquery-waypoints is
an optional one; as there is no way to differentiate hard and soft
dependencies in Bower, we're marking the latter as a "dev" dependency.

I've marked the jQuery dependency fairly loosely, to work with the
different versions of jQuery that we are using in multiple places (eg.
causes.git, which is on 1.10, and twigg.git, which is on 2.0). Likewise,
I've marked the jquery-waypoints dependency at 2.0.2, which is not the
latest (2.0.3 is), but will be satisified everywhere we use replace.js.

[0]: http://bower.io/
[1]: https://github.com/causes/method-proxy-js/commit/cc44dbc7934

Change-Id: I36bc11871016c38774998c51928983b493f0e93d
Reviewed-on: https://gerrit.causes.com/26425
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>